move half of runtime-utils to a new runtime-diff-utils
