+++ ui/src/app/components/runtime-configuration-panel.tsx
@@ -8,6 +8,11 @@ import {
   RuntimeStatus,
 } from 'generated/fetch';
 
+import {
+  AnalysisDiff,
+  diffsToUpdateMessaging,
+  getAnalysisConfigDiffs,
+} from '../utils/runtime-diff-utils';
 import { cond, switchCase } from '@terra-ui-packages/core-utils';
 import { Button } from 'app/components/buttons';
 import { Spinner } from 'app/components/spinners';
@@ -32,10 +37,7 @@ import { useCustomRuntime, useRuntimeStatus } from 'app/utils/runtime-hooks';
 import { applyPresetOverride } from 'app/utils/runtime-presets';
 import {
   AnalysisConfig,
-  AnalysisDiff,
-  diffsToUpdateMessaging,
   fromAnalysisConfig,
-  getAnalysisConfigDiffs,
   isVisible,
   maybeWithPersistentDisk,
   PanelContent,
