+++ ui/src/app/utils/runtime-utils.spec.tsx
@@ -19,20 +19,11 @@ import { render, waitFor } from '@testing-library/react';
 import { registerApiClient } from 'app/services/swagger-fetch-clients';
 import {
   AnalysisConfig,
-  AnalysisDiff,
-  AnalysisDiffState,
-  applyUpdate,
   canUseExistingDisk,
-  compareGpu,
-  diffsToUpdateMessaging,
   DiskConfig,
-  findMostSevereDiffState,
   fromAnalysisConfig,
   getCreator,
   maybeWithPersistentDisk,
-  rebootUpdate,
-  recreateEnvAndPDUpdate,
-  recreateEnvUpdate,
   toAnalysisConfig,
   withAnalysisConfigDefaults,
 } from 'app/utils/runtime-utils';
@@ -61,6 +52,12 @@ import {
   DEFAULT_DISK_SIZE,
   DEFAULT_MACHINE_TYPE,
 } from './machines';
+import {
+  AnalysisDiff,
+  AnalysisDiffState,
+  compareGpu,
+  diffsToUpdateMessaging,
+} from './runtime-diff-utils';
 import { useCustomRuntime } from './runtime-hooks';
 import { runtimePresets } from './runtime-presets';
 
@@ -152,47 +149,6 @@ describe('runtime-utils', () => {
   });
 });
 
-describe(findMostSevereDiffState.name, () => {
-  test.each([
-    [[], undefined],
-    [[AnalysisDiffState.NEEDS_DELETE], AnalysisDiffState.NEEDS_DELETE],
-    [
-      [AnalysisDiffState.NEEDS_DELETE, undefined],
-      AnalysisDiffState.NEEDS_DELETE,
-    ],
-    [
-      [
-        AnalysisDiffState.CAN_UPDATE_IN_PLACE,
-        AnalysisDiffState.NEEDS_DELETE,
-        AnalysisDiffState.NO_CHANGE,
-        AnalysisDiffState.CAN_UPDATE_WITH_REBOOT,
-        AnalysisDiffState.NO_CHANGE,
-      ],
-      AnalysisDiffState.NEEDS_DELETE,
-    ],
-    [
-      [
-        AnalysisDiffState.CAN_UPDATE_IN_PLACE,
-        AnalysisDiffState.NO_CHANGE,
-        AnalysisDiffState.CAN_UPDATE_WITH_REBOOT,
-        AnalysisDiffState.CAN_UPDATE_IN_PLACE,
-      ],
-      AnalysisDiffState.CAN_UPDATE_WITH_REBOOT,
-    ],
-    [
-      [
-        AnalysisDiffState.NO_CHANGE,
-        AnalysisDiffState.CAN_UPDATE_IN_PLACE,
-        AnalysisDiffState.NO_CHANGE,
-      ],
-      AnalysisDiffState.CAN_UPDATE_IN_PLACE,
-    ],
-    [[AnalysisDiffState.NO_CHANGE], AnalysisDiffState.NO_CHANGE],
-  ])('findMostSevereDiffState(%s) = %s', (diffStates, want) => {
-    expect(findMostSevereDiffState(diffStates)).toEqual(want);
-  });
-});
-
 describe(getCreator.name, () => {
   test.each([
     ['a runtime without a creator label', defaultRuntime(), undefined],
@@ -274,95 +230,6 @@ describe(maybeWithPersistentDisk.name, () => {
       disk.diskType
     );
   });
-});
-
-// diff state combination logic is covered by the findMostSevereDiffState tests
-// so here we can use single-element lists as inputs
-describe(diffsToUpdateMessaging.name, () => {
-  const baseDiff: AnalysisDiff = {
-    desc: undefined, // these are not used by this function
-    previous: undefined,
-    new: undefined,
-    diff: undefined, // tests will override
-  };
-
-  // diff = [NO_CHANGE, CAN_UPDATE_IN_PLACE] x diskDiff = [all states]
-  test.each([
-    [AnalysisDiffState.NO_CHANGE, AnalysisDiffState.NO_CHANGE],
-    [AnalysisDiffState.NO_CHANGE, AnalysisDiffState.CAN_UPDATE_IN_PLACE],
-    [AnalysisDiffState.NO_CHANGE, AnalysisDiffState.CAN_UPDATE_WITH_REBOOT],
-    [AnalysisDiffState.NO_CHANGE, AnalysisDiffState.NEEDS_DELETE],
-    [AnalysisDiffState.CAN_UPDATE_IN_PLACE, AnalysisDiffState.NO_CHANGE],
-    [
-      AnalysisDiffState.CAN_UPDATE_IN_PLACE,
-      AnalysisDiffState.CAN_UPDATE_IN_PLACE,
-    ],
-    [
-      AnalysisDiffState.CAN_UPDATE_IN_PLACE,
-      AnalysisDiffState.CAN_UPDATE_WITH_REBOOT,
-    ],
-    [AnalysisDiffState.CAN_UPDATE_IN_PLACE, AnalysisDiffState.NEEDS_DELETE],
-  ])(
-    'it returns an applyUpdate when the environment diff is %s and the diskDiff is %',
-    (diff, diskDiff) =>
-      expect(
-        diffsToUpdateMessaging([
-          {
-            ...baseDiff,
-            diff,
-            diskDiff,
-          },
-        ])
-      ).toEqual(applyUpdate)
-  );
-
-  test.each([
-    AnalysisDiffState.NO_CHANGE,
-    AnalysisDiffState.CAN_UPDATE_IN_PLACE,
-    AnalysisDiffState.CAN_UPDATE_WITH_REBOOT,
-    AnalysisDiffState.NEEDS_DELETE,
-  ])(
-    'it returns a rebootUpdate when the environment diff is CAN_UPDATE_WITH_REBOOT and the diskDiff is %s',
-    (diskDiff) =>
-      expect(
-        diffsToUpdateMessaging([
-          {
-            ...baseDiff,
-            diff: AnalysisDiffState.CAN_UPDATE_WITH_REBOOT,
-            diskDiff,
-          },
-        ])
-      ).toEqual(rebootUpdate)
-  );
-
-  test.each([
-    AnalysisDiffState.NO_CHANGE,
-    AnalysisDiffState.CAN_UPDATE_IN_PLACE,
-    AnalysisDiffState.CAN_UPDATE_WITH_REBOOT,
-  ])(
-    'it returns a recreateEnvUpdate when the environment diff is NEEDS_DELETE and the diskDiff is %s',
-    (diskDiff) =>
-      expect(
-        diffsToUpdateMessaging([
-          {
-            ...baseDiff,
-            diff: AnalysisDiffState.NEEDS_DELETE,
-            diskDiff,
-          },
-        ])
-      ).toEqual(recreateEnvUpdate)
-  );
-
-  it('returns a recreateEnvAndPDUpdate when both the environment diff and disk diff are NEEDS_DELETE', () =>
-    expect(
-      diffsToUpdateMessaging([
-        {
-          ...baseDiff,
-          diff: AnalysisDiffState.NEEDS_DELETE,
-          diskDiff: AnalysisDiffState.NEEDS_DELETE,
-        },
-      ])
-    ).toEqual(recreateEnvAndPDUpdate));
 });
 
 // only testing the 'diff' field because the others are straightforward
